requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
Pillow==11.2.1
requests==2.32.3
apscheduler==3.11.0
numpy==2.2.4
